requirements.txt
pytest==6.2.5
pdfminer.six==20201018
PyPDF2==1.26.0
pyyaml==5.3.1
requests==2.24.0
urllib3==1.25.11
pymongo==3.12.0